+++ shared/src/main/kotlin/ml/duncte123/skybot/database/PostgreDatabase.kt
@@ -30,7 +30,7 @@ import liquibase.database.jvm.JdbcConnection
 import liquibase.resource.ClassLoaderResourceAccessor
 import ml.duncte123.skybot.extensions.toGuildSetting
 import ml.duncte123.skybot.extensions.toReminder
-import ml.duncte123.skybot.extensions.toSQL
+import ml.duncte123.skybot.extensions.toSQLTimestamp
 import ml.duncte123.skybot.objects.Tag
 import ml.duncte123.skybot.objects.api.*
 import ml.duncte123.skybot.objects.command.CommandResult
@@ -670,13 +670,15 @@ class PostgreDatabase(jdbcURI: String, ohShitFn: (Int, Int) -> Unit = { _, _ -> 
                 }
             }
 
-            val values = warningIds.joinToString(", ") { "?" }
+            if (warningIds.isNotEmpty()) {
+                val values = warningIds.joinToString(", ") { "?" }
 
-            con.prepareStatement("DELETE FROM warnings WHERE id in ($values)").use { smt ->
-                warningIds.forEachIndexed { index, id ->
-                    smt.setInt(index + 1, id)
+                con.prepareStatement("DELETE FROM warnings WHERE id in ($values)").use { smt ->
+                    warningIds.forEachIndexed { index, id ->
+                        smt.setInt(index + 1, id)
+                    }
+                    smt.execute()
                 }
-                smt.execute()
             }
         }
 
@@ -937,7 +939,7 @@ class PostgreDatabase(jdbcURI: String, ohShitFn: (Int, Int) -> Unit = { _, _ -> 
                 smt.setLong(4, messageId)
                 smt.setBoolean(5, inChannel)
                 smt.setString(6, reminder)
-                smt.setDate(7, expireDate.toSQL())
+                smt.setTimestamp(7, expireDate.toSQLTimestamp())
 
                 try {
                     smt.execute()
